refactor: ActiveDecoratorを導入し、リファクタリングを行いました
