+++ Gemfile
@@ -61,6 +61,9 @@ gem "aws-sdk-s3", "~> 1.181"
 # Use Active Storage variants [https://guides.rubyonrails.org/active_storage_overview.html#transforming-images]
 gem "image_processing", "~> 1.2"
 
+# Use Active decorator
+gem 'active_decorator'
+
 
 group :development, :test do
   # See https://guides.rubyonrails.org/debugging_rails_applications.html#debugging-with-the-debug-gem
