+++ app/views/commons/_articles_tabs.html.haml
@@ -1,5 +1,5 @@
 .tabs
-  .tabs_item.active
+  .tabs_item{class: "#{add_active_class(root_path)}"}
     = link_to '新着記事一覧', root_path
-  .tabs_item
+  .tabs_item{class: "#{add_active_class(timeline_path)}"}
     = link_to 'タイムライン', timeline_path
